+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.119] - Not released
+### Added
+- Show the next comment in the same comment preview panel when clicking on the
+  "^^" in the comment preview.
 ### Changed
 - Use Vite as builder and Vitest as test runner. We no longer need Webpack,
   Mocha, Jest, and (mostly) Babel! Also: